editors - move some code out of common/editor.ts
